requirements.txt
python-telegram-bot
pillow
requests
aiohttp
redis